+++ src/types/module.ts
@@ -21,7 +21,8 @@ export interface ImageProviders {
   imgix?: any
   prismic?: any
   twicpics?: any
-  storyblok?: any,
+  storyblok?: any
+  imageengine?: any,
   ipx?: Partial<IPXOptions>
   static?: Partial<IPXOptions>
 }