+++ src/module.ts
@@ -17,6 +17,7 @@ const imageModule: Module<ModuleOptions> = async function imageModule (moduleOpt
     dir: resolve(nuxt.options.srcDir, nuxt.options.dir.static),
     domains: [],
     sharp: {},
+    strategy: 'desktop-first',
     // https://tailwindcss.com/docs/breakpoints
     screens: {
       xs: 320,
@@ -51,7 +52,8 @@ const imageModule: Module<ModuleOptions> = async function imageModule (moduleOpt
     'presets',
     'provider',
     'domains',
-    'alias'
+    'alias',
+    'strategy'
   ])
 
   const providers = resolveProviders(nuxt, options)