+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.7.0](https://github.com/nuxt/image/compare/v0.6.1...v0.7.0) (2022-02-03)
+
+
+### Features
+
+* add `strategy` property to config in order to have a mobile-first strategy rather than the default desktop-first ([#360](https://github.com/nuxt/image/issues/360))
+
 ### [0.6.1](https://github.com/nuxt/image/compare/v0.6.0...v0.6.1) (2022-01-31)